added messagebox to make sure of deletion

+++ Alligator/Commands/TabItemClients/DeleteClient_AllClientsPage.cs
@@ -23,13 +23,16 @@ namespace Alligator.UI.Commands.TabItemClients
             }
             else
             {
-                //TODO: messagebox to make sure of deletion
+                
                 //TODO: сделать работу с базой через try catch
                 //TODO: отлавливать ошибки, если вдруг есть связи и нельзя удалить клиента()(())(
-
-                _commentService.DeleteCommentsByClientId(viewModel.SelectedClient.Id);
-                _clientService.DeleteClient(viewModel.SelectedClient);
-                viewModel.Clients.Remove(viewModel.SelectedClient);
+                var userAnswer = MessageBox.Show("Вы правда хотите удалить этого клиента?", "Удаление", MessageBoxButton.YesNo, MessageBoxImage.Question);
+                if (userAnswer == MessageBoxResult.Yes)
+                {
+                    _commentService.DeleteCommentsByClientId(viewModel.SelectedClient.Id);
+                    _clientService.DeleteClient(viewModel.SelectedClient);
+                    viewModel.Clients.Remove(viewModel.SelectedClient);
+                }
             }
         }
     }
